refactor(core): decompose Sonos client into focused modules
Split the 1,280-line client.rs into submodules by UPnP service:
- zone_groups.rs: topology parsing and retrieval
- didl.rs: DIDL-Lite metadata formatting
- playback.rs: play/stop/transport control with retry logic
- volume.rs: group and per-speaker volume/mute
- grouping.rs: join/leave group coordination

client.rs now contains only SonosClientImpl and trait delegation.
