+++ packages/thaumic-core/src/sonos/mod.rs
@@ -8,7 +8,12 @@
 //! - `types` - Domain types for zone groups and speakers
 //! - `services` - UPnP service definitions (URNs, paths)
 //! - `traits` - Trait abstractions for testability
-//! - `client` - High-level Sonos commands (play, stop, volume, mute, zone groups)
+//! - `client` - `SonosClientImpl` concrete trait implementation
+//! - `zone_groups` - Zone group topology parsing and retrieval
+//! - `didl` - DIDL-Lite metadata formatting for Sonos display
+//! - `playback` - Play, stop, and transport control commands
+//! - `volume` - Group and per-speaker volume/mute control
+//! - `grouping` - Group join/leave coordination
 //! - `discovery` - Multi-method speaker discovery (SSDP multicast/broadcast + mDNS)
 //! - `gena` - UPnP GENA event subscription lifecycle (coordinator)
 //! - `gena_client` - GENA HTTP operations
@@ -19,18 +24,23 @@
 //! - `utils` - Shared utility functions
 
 pub mod client;
+pub(crate) mod didl;
 pub mod discovery;
 pub mod gena;
 pub mod gena_client;
 pub mod gena_event_builder;
 pub mod gena_parser;
 pub mod gena_store;
+pub mod grouping;
+pub mod playback;
 pub mod services;
 pub mod soap;
 pub mod subscription_arbiter;
 pub mod traits;
 pub mod types;
 pub mod utils;
+pub mod volume;
+pub mod zone_groups;
 
 #[cfg(test)]
 pub(crate) mod test_fixtures;
